refactor: cogs and prep for adding rss feeds
