+++ feed_bot/utils/reddit.py
@@ -5,16 +5,14 @@ from aiohttp import ClientSession
 import asyncpraw
 from asyncprawcore.exceptions import ResponseException
 
+from .common import CommonUtilities
 
-class Reddit:
-    """Reddit
+
+class Reddit(CommonUtilities):
+    """Utility class for interacting with the reddit api
 
     Reddit API: https://www.reddit.com/dev/api/
     """
-
-    error = False
-    error_msg = ""
-    res_dicts = []
 
     def __init__(
         self,
@@ -22,8 +20,8 @@ class Reddit:
         subreddit_names: [str] = "",
         channel_id: str = "",
     ) -> None:
+        super().__init__(session=session, channel_id=channel_id)
         self.subreddits_query = "+".join(subreddit_names)
-        self.channel_id = channel_id
         self.reddit = asyncpraw.Reddit(
             client_id=os.getenv("REDDIT_CLIENT_ID"),
             client_secret=os.getenv("REDDIT_CLIENT_SECRET"),
@@ -32,11 +30,6 @@ class Reddit:
             user_agent=os.getenv("REDDIT_USER_AGENT"),
             username=os.getenv("REDDIT_USERNAME"),
         )
-
-    def clear(self):
-        self.error = False
-        self.error_msg = ""
-        self.res_dicts = []
 
     async def get_subreddit_submissions(self, *args, **kwargs) -> None:
         self.clear()
@@ -60,7 +53,7 @@ class Reddit:
                     self.res_dicts.append(submission_dict)
 
     @staticmethod
-    def documents_to_embeds(documents, *args, **kwargs):
+    def documents_to_embeds(documents: [dict], *args, **kwargs):
         """Static method for converting noSql Documents to Discord Embeds"""
         channel_embeds = []
         for doc in documents:
