+++ feed_bot/utils/rss.py
@@ -1,0 +1,8 @@
+import feedparser
+from .common import CommonUtilities
+
+
+class WebFeed(CommonUtilities):
+    """Utility class for interacting with web rss feeds"""
+
+    pass
